Closes VIZ-522: crash when column is unknown in settings
