+++ frontend/src/metabase/visualizer/utils/index.ts
@@ -2,6 +2,7 @@ export * from "./column";
 export * from "./dashboard-card-supports-visualizer";
 export * from "./data-source";
 export * from "./drag-and-drop";
+export * from "./get-initial-state-for-card-data-source";
 export * from "./is-visualizer-dashboard-card";
 export * from "./merge-data";
 export * from "./misc";
