+++ frontend/src/metabase/visualizer/utils/misc.ts
@@ -1,7 +1,5 @@
 import { t } from "ttag";
 
-import { isNotNull } from "metabase/lib/types";
-import { getColumnVizSettings } from "metabase/visualizations";
 import { isDate, isNumeric } from "metabase-lib/v1/types/utils/isa";
 import type {
   Card,
@@ -9,14 +7,6 @@ import type {
   VisualizationDisplay,
   VisualizationSettings,
 } from "metabase-types/api";
-import type { VisualizerHistoryItem } from "metabase-types/store/visualizer";
-
-import {
-  copyColumn,
-  createVisualizerColumnReference,
-  extractReferencedColumns,
-} from "./column";
-import { createDataSource } from "./data-source";
 
 export function getDefaultVisualizationName() {
   return t`My new visualization`;
@@ -83,59 +73,4 @@ function areAreaBarLineSeriesCompatible(
     (isDate(primaryInitialDimension) === isDate(primaryNewDimension) ||
       isNumeric(primaryInitialDimension) !== isNumeric(primaryNewDimension))
   );
-}
-
-export function getInitialStateForCardDataSource(
-  card: Card,
-  columns: DatasetColumn[],
-): VisualizerHistoryItem {
-  const state: VisualizerHistoryItem = {
-    display: card.display,
-    columns: [],
-    columnValuesMapping: {},
-    settings: {},
-  };
-  const dataSource = createDataSource("card", card.id, card.name);
-
-  columns.forEach(column => {
-    const columnRef = createVisualizerColumnReference(
-      dataSource,
-      column,
-      extractReferencedColumns(state.columnValuesMapping),
-    );
-    state.columns.push(copyColumn(columnRef.name, column));
-    state.columnValuesMapping[columnRef.name] = [columnRef];
-  });
-
-  const entries = getColumnVizSettings(card.display)
-    .map(setting => {
-      const originalValue = card.visualization_settings[setting];
-
-      if (!originalValue) {
-        return null;
-      }
-
-      if (Array.isArray(originalValue)) {
-        return [
-          setting,
-          originalValue.map(originalColumnName => {
-            const index = columns.findIndex(
-              col => col.name === originalColumnName,
-            );
-            return state.columns[index].name;
-          }),
-        ];
-      } else {
-        const index = columns.findIndex(col => col.name === originalValue);
-        return [setting, state.columns[index].name];
-      }
-    })
-    .filter(isNotNull);
-
-  state.settings = {
-    ...card.visualization_settings,
-    ...Object.fromEntries(entries),
-  };
-
-  return state;
 }
